Material from Recipes
Added tooltips to show the material of the tool you are holding.
Added "Developer Mode" config option. When using this option, the exact material id will show as a tooltip on tools.
Added "armory_recipe" recipes. These allow you to craft tools with specific materials on the output. They only have one input, it goes in the top center of the crafting table.